+++ packages/editor/src/bindings/post-entity/index.js
